refactor(ui): разделена presentation-логика SettingsScreen
root SettingsScreenController оставлен координатором loading/apply flow и dispatch пользовательских действий
section builders и action handlers разнесены по settings areas
