+++ RSSReader/Views/SettingsScreen/SettingsScreenController+Notifications.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+@MainActor
+extension SettingsScreenController {
+    func updateShowUnreadCountBadge(
+        isOn: Bool,
+        dependencies: AppDependencies
+    ) {
+        guard screenState.settingsInput.showUnreadCountBadge != isOn else {
+            return
+        }
+
+        var input = screenState.settingsInput
+        input.showUnreadCountBadge = isOn
+        screenState.applyDraftInput(input)
+    }
+}
